Moved search files to folder
